move ListModels to CommonEditor for re-use
